modularized actions and extended test coverage
