+++ src/com/twitter/intellij/pants/components/impl/PantsInitComponentImpl.java
@@ -12,7 +12,6 @@ import com.intellij.notification.NotificationType;
 import com.intellij.notification.Notifications;
 import com.intellij.openapi.actionSystem.ActionManager;
 import com.intellij.openapi.actionSystem.AnAction;
-import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.IdeActions;
 import com.intellij.openapi.actionSystem.KeyboardShortcut;
 import com.intellij.openapi.extensions.PluginId;
@@ -20,20 +19,15 @@ import com.intellij.openapi.externalSystem.util.ExternalSystemConstants;
 import com.intellij.openapi.keymap.Keymap;
 import com.intellij.openapi.keymap.KeymapManager;
 import com.intellij.openapi.options.ShowSettingsUtil;
-import com.intellij.openapi.project.DumbAware;
-import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.registry.Registry;
 import com.intellij.openapi.util.text.StringUtil;
 import com.twitter.intellij.pants.components.PantsInitComponent;
 import com.twitter.intellij.pants.ui.PantsCompileAllTargetsAction;
 import com.twitter.intellij.pants.ui.PantsCompileAllTargetsInModuleAction;
+import com.twitter.intellij.pants.ui.PantsOverrideAction;
 import com.twitter.intellij.pants.ui.PantsRebuildAction;
 import com.twitter.intellij.pants.util.PantsConstants;
-import com.twitter.intellij.pants.util.PantsUtil;
-import icons.PantsIcons;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.jetbrains.annotations.TestOnly;
 
 import javax.swing.event.HyperlinkEvent;
 import java.io.File;
@@ -106,102 +100,6 @@ public class PantsInitComponentImpl implements PantsInitComponent {
     actionManager.registerAction(IdeActions.ACTION_COMPILE_PROJECT, pantsRebuildAction);
   }
 
-  //  Used to toggle between two actions, one that is active for Pants projects, and
-  //  another for other projects.
-  private static class PantsOverrideAction extends AnAction implements DumbAware {
-
-    private AnAction secondaryIdeaAction;
-    private AnAction primaryPantsAction;
-    private boolean pantsActive;
-
-    private PantsOverrideAction(String actionId, @NotNull AnAction pantsAction) {
-      basicConstruction(actionId);
-      primaryPantsAction = pantsAction;
-    }
-
-    private PantsOverrideAction(String actionId, String oldName, @NotNull AnAction pantsAction) {
-      super(oldName);
-      basicConstruction(actionId);
-      primaryPantsAction = pantsAction;
-    }
-
-
-    private void basicConstruction(String actionId) {
-      secondaryIdeaAction = getAction(actionId);
-      primaryPantsAction = new PantsShieldAction(secondaryIdeaAction);
-    }
-
-    private static PantsOverrideAction createDisabledEmptyAction(String actionId) {
-      return new PantsOverrideAction(actionId, new PantsShieldAction(getAction(actionId)));
-    }
-
-    private boolean isPantsProject(AnActionEvent event) {
-      if (event == null) {
-        return false;
-      }
-      Project project = event.getProject();
-      final boolean result = project != null && PantsUtil.isPantsProject(project);
-      pantsActive = result;
-      return pantsActive;
-    }
-
-    @Override
-    public void update(AnActionEvent event) {
-      if (secondaryIdeaAction != null) {
-        secondaryIdeaAction.update(event);
-      }
-      if (isPantsProject(event)) {
-        event.getPresentation().setIcon(PantsIcons.Icon);
-        primaryPantsAction.update(event);
-      }
-    }
-
-    @Override
-    public void actionPerformed(AnActionEvent event) {
-      if (isPantsProject(event)) {
-        primaryPantsAction.actionPerformed(event);
-      }
-      else if (secondaryIdeaAction != null) {
-        secondaryIdeaAction.actionPerformed(event);
-      }
-    }
-
-    @Override
-    @TestOnly
-    public String toString() {
-      String activeOverride = pantsActive ? " actively" : "";
-      return
-        primaryPantsAction.getClass().getSimpleName() + activeOverride +  " overriding " +
-        (secondaryIdeaAction == null ? "NullAction" : secondaryIdeaAction.getClass().getSimpleName());
-    }
-  }
-
-  /**
-   * Shield action allows previous action's text to still be displayed
-   * and update, but disables all possible user interaction
-   */
-  private static class PantsShieldAction extends AnAction {
-    private AnAction shieldedAction;
-
-    private PantsShieldAction(@Nullable AnAction action) {
-      shieldedAction = action;
-    }
-
-    @Override
-    public void actionPerformed(AnActionEvent event) {}
-
-    @Override
-    public void update(AnActionEvent event) {
-      if (shieldedAction != null) {
-        shieldedAction.update(event);
-      }
-      event.getPresentation().setEnabled(false);
-    }
-  }
-
-  private static AnAction getAction(@NotNull String actionId) {
-    return ActionManager.getInstance().getAction(actionId);
-  }
 
   private void registerRefreshKey() {
     Keymap keymap = KeymapManager.getInstance().getActiveKeymap();
